추가 코드 프로그램 실행은 가능함 하지만 의도한 대로의 랜더링 결과는 아님 반사하는 부분이 이상하게 보이는 경우가 있음 이대로 commit 한 이유는 엔진 코드 자체를 대대적으로 바꿀 예정 이대로 진행하면 엔진 구조를 바꿀지 못하고, 나중에 content 만들 때, 많은 문제가 생길 것 같음, content 수정은 이정도로 마무리 하고 엔진 구조를 수정하는 쪽이 낮다고 판단함

+++ app/component/VKGeometryGenerator.cpp
@@ -3,7 +3,7 @@
 namespace vkengine
 {
     namespace helper {
-        void GeometryGenerator::createSphere(std::vector<Vertex>& vertices, std::vector<uint16_t>& indices, float radius, uint32_t rings, uint32_t sectors)
+        void GeometryGenerator::createSphere(std::vector<Vertex>& vertices, std::vector<uint32_t>& indices, float radius, uint32_t rings, uint32_t sectors)
         {
             vertices.clear();
             indices.clear();
@@ -60,7 +60,7 @@ namespace vkengine
             }
 
             // ÀÎµ¦½º »ý¼º
-            uint16_t k1, k2;
+            uint32_t k1, k2;
             for (int i = 0; i < sectors; ++i)
             {
                 k1 = i * (rings + 1);     // beginning of current stack
